add upgrade markers and docs
